#65 - fix route room id not correct when join room and optimize socketio connnect from client to server
